%% ----------------------------------------------------------------------------
%% Header
\name{Kapitel 0}
\alias{Kapitel 0}
\encoding{UTF-8}
\title{Kapitel 0: Testkonzeption: das Konzept der österreichischen 
Bildungsstandards-Überprüfung}

\description{
Das ist die Nutzerseite zum Kapitel 0, Testkonzeption: das Konzept der 
österreichischen Bildungsstandards-Überprüfung, im Herausgeberband 
\emph{Large-Scale Assessment mit \R}:  Methodische Grundlagen der österreichischen 
Bildungsstandard-Überprüfung. 
Hier werden die im Kapitel verwendeten \R-Syntaxen zur Unterstützung für 
Leser/innen kommentiert, dokumentiert und gegebenefalls erweitert.
}

%% ----------------------------------------------------------------------------
%% Intermediate Deklarationes
\references{
  Breit, S., Schreiner, C. & Fellinger, R. (2016). Testkonzeption: das Konzept 
  der österreichischen Bildungsstandards-Überprüfung. In C. Schreiner 
  & S. Breit (Hrsg.): \emph{Large-Scale Assessment mit \R}:  Methodische Grundlagen 
  der österreichischen Bildungsstandard-Überprüfung (pp. xx-yy). Wien: facultas
  wuv.
}

% \note{
% % wenn es Hinweise gibt
% }

\author{
  Simone Breit, Claudia Schreiner und Robert Fellinger
}

\seealso{
Zu \code{\link{Kapitel 1}}, Testkonstruktion.\cr
Zur \code{\link[LSAmitR:LSAmitR-package]{Übersicht}}.
%Verweis zu Hilfeseiten eventuell zusätzlicher R-Funktionen. (\cr nach Übersicht einfügen)
}

%% ----------------------------------------------------------------------------
%% Details
\details{
Dieses Kapitel enthält keine Beispiele mit \R.
}

%% ----------------------------------------------------------------------------
%% Examples
%\examples{
%##
%}