%% ----------------------------------------------------------------------------
%% Header
\name{Kapitel 11}
\alias{Kapitel 11}
\encoding{UTF-8}
\title{Kapitel 11: Aspekte der Validierung}

\description{
Das ist die Nutzerseite zum Kapitel 11, Aspekte der Validierung, im 
Herausgeberband \emph{Large-Scale Assessment mit \R}:  Methodische Grundlagen 
der österreichischen Bildungsstandard-Überprüfung. 
Hier werden die im Kapitel verwendeten \R-Syntaxen zur Unterstützung für 
Leser/innen kommentiert, dokumentiert und gegebenefalls erweitert.
}

%% ----------------------------------------------------------------------------
%% Intermediate Deklarationes
\references{
  Fellinger, R., Kiefer, T., Robitzsch, A. & Trendtel, M. (2016). Aspekte der
  Validierung. In C. Schreiner & S. Breit (Hrsg.): \emph{Large-Scale Assessment 
  mit \R}:  Methodische Grundlagen der österreichischen 
  Bildungsstandard-Überprüfung (pp. xx-yy). Wien: facultas wuv.
}

% \note{
% % wenn es Hinweise gibt
% }

\author{
  Robert Fellinger, Thomas Kiefer, Alexander Robitzsch, Matthias Trendtel
}

\seealso{
Zurück zu \code{\link{Kapitel 10}}, Reporting und Analysen.\cr
Zur \code{\link[LSAmitR:LSAmitR-package]{Übersicht}}.
%Verweis zu Hilfeseiten eventuell zusätzlicher R-Funktionen. (\cr nach Übersicht einfügen)
}

%% ----------------------------------------------------------------------------
%% Details
\details{
Dieses Kapitel enthält keine Beispiele mit \R.
}

%% ----------------------------------------------------------------------------
%% Examples
%\examples{
%##
%}