%% ----------------------------------------------------------------------------
%% Header
\name{Kapitel 11}
\alias{Kapitel 11}
\encoding{UTF-8}
\title{Kapitel 11: Aspekte der Validierung}

\description{
Das ist die Nutzerseite zum Kapitel 11, \emph{Aspekte der Validierung}, im 
Herausgeberband Large-Scale Assessment mit \R: Methodische 
Grundlagen der österreichischen Bildungsstandardüberprüfung. 
Im Abschnitt \strong{Details} werden die im Kapitel verwendeten \R-Syntaxen zur 
Unterstützung für Leser/innen kommentiert und dokumentiert. 
Im Abschnitt \strong{Examples} werden die \R-Syntaxen des Kapitels vollständig 
wiedergegeben und gegebenenfalls erweitert.
}

%% ----------------------------------------------------------------------------
%% Intermediate Deklarationes
\references{
  Fellinger, R., Kiefer, T., Robitzsch, A. & Trendtel, M. (2016). Aspekte der
  Validierung.
  In S. Breit & C. Schreiner (Hrsg.), \emph{Large-Scale Assessment mit \R:  
  Methodische Grundlagen der österreichischen Bildungsstandardüberprüfung} 
  (pp. 363--398). Wien: facultas.
}

% \note{
% % wenn es Hinweise gibt
% }

\author{
  Robert Fellinger, Thomas Kiefer, Alexander Robitzsch, Matthias Trendtel
}

\seealso{
Zurück zu \code{\link{Kapitel 10}}, Reporting und Analysen.\cr
Zur \code{\link[LSAmitR:LSAmitR-package]{Übersicht}}.
%Verweis zu Hilfeseiten eventuell zusätzlicher R-Funktionen. (\cr nach 
%Übersicht einfügen)
}

%% ----------------------------------------------------------------------------
%% Details
\details{
Dieses Kapitel enthält keine Beispiele mit \R.
}

%% ----------------------------------------------------------------------------
%% Examples
%\examples{
%##
%}