\name{datenKapitel01}
\alias{datenKapitel01}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 1
}


\description{
...
}


\usage{
data(datenKapitel01)
}


\format{
\code{datenKapitel01} ist eine Liste mit vier Elementen
\itemize{
\item \code{datenKapitel01$pilotScored} 
\verb{
'data.frame':	2504 obs. of  163 variables:
 $ sidstud  : int  1052 1057 1058 1064 1068 1073 1074 1076 1078 1080 ...
 $ female   : chr  "w" "w" "w" "w" ...
 $ form     : chr  "PR019" "PR020" "PR021" "PR022" ...
 $ E8RS13151: int  NA NA NA NA NA NA NA NA NA NA ...
 $ E8RS13171: int  NA NA 1 NA NA NA NA 1 NA NA ...
 $ E8RS13491: int  NA NA NA NA 0 NA NA NA NA NA ...
 $ E8RS13641: int  0 NA NA NA NA NA 0 NA NA NA ...
 [...] 
 $ E8RS7929: int  NA NA 0 NA NA NA NA NA NA NA ...
 $ E8RS7940: int  NA NA NA NA NA 0 NA NA NA NA ...
 $ E8RS7955: int  NA 0 NA NA NA NA NA NA 2 NA ...
 $ E8RS7993: int  NA NA NA 0 NA NA 2 NA NA NA ...
}

\item \code{datenKapitel01$pilotScored} 
\verb{
'data.frame':	320 obs. of  65 variables:
 $ testlet   : chr  "E8RS1315" "E8RS1317" "E8RS1340" "E8RS1349" ...
 $ task      : chr  "E8RS13151" "E8RS13171" "E8RS13401" "E8RS13491" ...
 $ format    : chr  "MC3" "MC3" "MC3" "MC3" ...
 $ focus     : int  1 1 1 1 1 1 1 1 1 1 ...
 $ focusLabel: chr  "RFocus1" "RFocus1" "RFocus1" "RFocus1" ...
 $ topic     : chr  "Interkulturelle und landeskundliche Aspekte" "Familie und Freunde" "Umgang mit Geld" "Einstellungen und Werte" ...
 $ no.Words  : int  24 24 29 32 10 33 22 41 10 37 ...
 $ key       : chr  "1" "3" "2" "2" ...
 $ maxScore  : int  1 1 1 1 1 1 1 1 1 1 ...
 $ PR001     : int  NA NA NA 10 NA NA NA NA NA NA ...
 $ PR002     : int  5 NA 6 NA 7 NA NA 8 NA NA ...
 $ PR003     : int  NA NA NA 6 NA NA NA NA NA NA ...
 $ PR004     : int  NA NA NA 10 NA NA NA NA NA NA ...
 [...]
 $ PR054     : int  NA NA NA NA NA NA NA NA NA NA ...
 $ PR055     : int  NA 9 NA NA NA NA 10 NA NA 11 ...
 $ PR056     : int  NA NA NA NA NA NA NA NA 6 NA ...
}

\item \code{datenKapitel01$pilotScored} 
\verb{
'data.frame':	2504 obs. of  323 variables:
 $ sidstud  : int  1052 1057 1058 1064 1068 1073 1074 1076 1078 1080 ...
 $ female   : chr  "w" "w" "w" "w" ...
 $ form     : chr  "PR019" "PR020" "PR021" "PR022" ...
 $ E8RS13151: int  NA NA NA NA NA NA NA NA NA NA ...
 $ E8RS13171: int  NA NA 3 NA NA NA NA 3 NA NA ...
 $ E8RS13491: int  NA NA NA NA 3 NA NA NA NA NA ...
 $ E8RS13641: int  2 NA NA NA NA NA 2 NA NA NA ...
 [...]
 $ E8RS7929: int  NA NA 0 NA NA NA NA NA NA NA ...
 $ E8RS7940: int  NA NA NA NA NA 0 NA NA NA NA ...
 $ E8RS7955: int  NA 0 NA NA NA NA NA NA 2 NA ...
 $ E8RS7993: int  NA NA NA 0 NA NA 2 NA NA NA ...
}

\item \code{datenKapitel01$pilotMM} 
\verb{
'data.frame':	1200 obs. of  5 variables:
 $ sidstud: int  1185 1269 1311 1522 1658 1665 1854 1889 1921 2067 ...
 $ task   : chr  "E8RS46051" "E8RS46051" "E8RS46051" "E8RS46051" ...
 $ Coder_1: int  1 1 9 0 0 9 9 1 9 0 ...
 $ Coder_2: int  1 1 9 0 0 9 9 1 9 0 ...
 $ Coder_3: int  1 1 9 0 0 9 9 1 9 0 ...
}

}
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
