\name{datenKapitel02}
\alias{datenKapitel02}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 2
}


\description{
...
}


\usage{
data(datenKapitel02)
}


\format{
\code{datenKapitel02} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
