\name{datenKapitel03}
\alias{datenKapitel03}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 3
}


\description{
...
}


\usage{
data(datenKapitel03)
}


\format{
\code{datenKapitel03} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
