\name{datenKapitel04}
\alias{datenKapitel04}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 4
}


\description{
...
}


\usage{
data(datenKapitel04)
}


\format{
\code{datenKapitel04} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
