\name{datenKapitel05}
\alias{datenKapitel05}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 5
}


\description{
...
}


\usage{
data(datenKapitel05)
}


\format{
\code{datenKapitel05} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
