\name{datenKapitel06}
\alias{datenKapitel06}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 6
}


\description{
...
}


\usage{
data(datenKapitel06)
}


\format{
\code{datenKapitel06} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
