\name{datenKapitel07}
\alias{datenKapitel07}
\alias{datenKapitel07Ex}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 7
}


\description{
...
}


\usage{
data(datenKapitel07)
}


\format{
\code{datenKapitel07} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
