\name{datenKapitel08}
\alias{datenKapitel08}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 8
}


\description{
...
}


\usage{
data(datenKapitel08)
}


\format{
\code{datenKapitel08} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
