\name{datenKapitel10}
\alias{datenKapitel10}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 10
}


\description{
...
}


\usage{
data(datenKapitel10)
}


\format{
\code{datenKapitel10} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
