%% ----------------------------------------------------------------------------
%% Header
\name{LSAmitR-Hilfsmethoden}

\alias{quintfunct}
\alias{covainteraction}
\alias{summary.VarComp}
\alias{zones.within.stratum}

\encoding{UTF-8}

\title{
Large-Scale Assessment mit R: Hilfsfunktionen aus den Kapiteln
}
\description{
Das ist die Nutzerseite zu den Hilfsfunktionen, die in einigen Kapiteln im 
Herausgeberband Large-Scale Assessment mit R: Methodische Grundlagen der 
österreichischen Bildungsstandard-Überprüfung angewendet werden.
}

%% ----------------------------------------------------------------------------
%% Intermediate Deklarationes
\usage{
zones.within.stratum(offset, n.str)

covainteraction(dat,covas, nchar)

quintfunct(X,w)
}

\arguments{
  \item{offset}{to fill}
  \item{n.str}{to fill}
  \item{dat}{to fill}
  \item{covas}{to fill}
  \item{nchar}{to fill}
  \item{X}{to fill}
  \item{w}{to fill}
}

% \references{
%
% }

% \note{
% wenn es Hinweise gibt
% }


% \seealso{
%Zur Hilfeseite von \code{\link[TAM:TAM-package]{TAM}}
% }


%% ----------------------------------------------------------------------------
%% Examples
% \examples{
% }