%% ----------------------------------------------------------------------------
%% Header
\name{Kapitel  0}
\alias{Kapitel 0}
\encoding{UTF-8}
\title{Kapitel 0: Konzeption der Ueberpruefung der Bildungsstandards in 
Oesterreich}

\description{
Das ist die Nutzerseite zum Kapitel 0, \emph{Konzeption der Überprüfung der 
Bildungsstandards in Österreich}, im Herausgeberband Large-Scale Assessment mit 
\R: Methodische Grundlagen der österreichischen 
Bildungsstandardüberprüfung. 
Hier werden die im Kapitel verwendeten \R-Syntaxen zur Unterstützung für 
Leser/innen kommentiert, dokumentiert und gegebenefalls erweitert.
}

%% ----------------------------------------------------------------------------
%% Intermediate Deklarationes
\references{
  Schreiner, C. & Breit, S. (2016). Konzeption der Überprüfung 
  der Bildungsstandards in Österreich.   
  In S. Breit & C. Schreiner (Hrsg.), \emph{Large-Scale Assessment mit \R:  
  Methodische Grundlagen der österreichischen Bildungsstandardüberprüfung} 
  (pp. 1--20). Wien: facultas.
}

% \note{
% % wenn es Hinweise gibt
% }

\author{
  Claudia Schreiner und Simone Breit
}

\seealso{
Zu \code{\link{Kapitel 1}}, Testkonstruktion.\cr
Zur \code{\link[LSAmitR:LSAmitR-package]{Übersicht}}.
%Verweis zu Hilfeseiten eventuell zusätzlicher R-Funktionen. (\cr nach Übersicht einfügen)
}

%% ----------------------------------------------------------------------------
%% Details
\details{
Dieses Kapitel enthält keine Beispiele mit \R.
}

%% ----------------------------------------------------------------------------
%% Examples
%\examples{
%##
%}