\name{datenKapitel09}
\alias{datenKapitel09}
\docType{data}
\encoding{UTF-8}

\title{
Daten Kapitel 9
}


\description{
...
}


\usage{
data(datenKapitel09)
}


\format{
\code{datenKapitel09} ist eine Liste mit vier Elementen
}

\seealso{
...
}	

\references{
...
}


\examples{
5+5
}

\keyword{datasets}
