%% ----------------------------------------------------------------------------
%% Header
\name{LSAmitR-Hilfsmethoden}

\alias{quintfunct}
\alias{covainteraction}
\alias{summary.VarComp}
\alias{zones.within.stratum}

\encoding{UTF-8}

\title{
Large-Scale Assessment mit R: Hilfsfunktionen aus den Kapiteln
}
\description{
Das ist die Nutzerseite zu den Hilfsfunktionen, die in einigen Kapiteln im 
Herausgeberband Large-Scale Assessment mit \R: Methodische 
Grundlagen der österreichischen Bildungsstandardüberprüfung angewendet werden.
}

%% ----------------------------------------------------------------------------
%% Intermediate Deklarationes
\usage{
zones.within.stratum(offset, n.str)

covainteraction(dat,covas, nchar)

quintfunct(X,w)
}

\arguments{
  \item{offset}{siehe \code{\link{Kapitel 2}}}
  \item{n.str}{siehe \code{\link{Kapitel 2}}}
  \item{dat}{siehe \code{\link{Kapitel 9}}}
  \item{covas}{siehe \code{\link{Kapitel 9}}}
  \item{nchar}{siehe \code{\link{Kapitel 9}}}
  \item{X}{\code{quintfunct} ist eine Hilfsfunktion, die nicht für die weitere 
  Verwendung gedacht ist.}
  \item{w}{\code{quintfunct} ist eine Hilfsfunktion, die nicht für die weitere 
  Verwendung gedacht ist.}
}

% \references{
%
% }

% \note{
% wenn es Hinweise gibt
% }


% \seealso{
%Zur Hilfeseite von \code{\link[TAM:TAM-package]{TAM}}
% }


%% ----------------------------------------------------------------------------
%% Examples
% \examples{
% }